+++ player.py
@@ -22,6 +22,7 @@ from json import dump
 
 import re
 
+from FinalHandler import FinalHandler
 from SignHandler import SignHandler
 from ViewTrack import ViewTrack
 from configs import config
@@ -40,7 +41,7 @@ class MainWindow(QMainWindow):
     def __init__(self):
         super().__init__()
         self.converter = Converter()
-
+        self.finalHandler = None
         self.calculation = None
         self.k = -0.5
         self.b = 10.9
@@ -185,6 +186,7 @@ class MainWindow(QMainWindow):
         self.button_viewTrack.setEnabled(is_filled_all)
         if is_filled_all:
             self.ViewTrack = ViewTrack()
+            self.finalHandler = FinalHandler()
 
     def set_path_to_video(self):
         config.PATH_TO_GPX = r"D:\Urban\vid\test\07,07,20211.gpx"
@@ -340,157 +342,26 @@ class MainWindow(QMainWindow):
         options |= QFileDialog.DontUseNativeDialog
         save_path, _ = QFileDialog.getSaveFileName(self, "QFileDialog.getSaveFileName()", "",
                                                    "geojson (*.geojson)", options=options)
-        save_path += '.geojson'
+        if save_path not in ".geojson":
+            save_path += ".geojson"
         config.PATH_TO_GEOJSON = save_path
         self.label_geojson.setText("{}".format(config.PATH_TO_GEOJSON))
         self.check_for_filling_of_data()
 
-    #TODO Вынести в отельный класс
-    # TODO handling_signs и handling_side объеденить в один метод
 
-    def handling_signs(self):
-        grouped_objects = {}
-        features = []
-        # Обработка результатов знаков
-
-        for obj in self.view.sign_handler.result_signs:
-            key = str(obj.car_coordinates_x[-1]) + str(obj.is_left)
-            grouped_objects.setdefault(key, []).append(obj)
-        for key, items in grouped_objects.items():
-            coefficient = 2
-            for item in items:
-                x1, y1, x2, y2 = self.calculation.get_line(item, coefficient)
-                coefficient += 1
-
-                feature = self.calculation.create_feature_object(x1, x2, y1, y2, item)
-                features.append(feature)
-        return features
-
-    def handling_side(self):
-        grouped_objects = {}
-        features = []
-        for obj in self.view.sign_handler.side_signs:
-            key = str(obj.car_coordinates_x[-1]) + str(obj.is_left)
-            grouped_objects.setdefault(key, []).append(obj)
-        for key, items in grouped_objects.items():
-            coefficient = 2
-            for item in items:
-                x1, y1, x2, y2 = self.calculation.get_line(item, coefficient)
-                coefficient += 1
-                item.azimuth = (item.azimuth + 90) % 360
-                feature = self.calculation.create_feature_object(x1, x2, y1, y2, item)
-                features.append(feature)
-        return features
-
-    def handling_turns(self):
-        features = []
-        # Обработка поворотов
-        for turn in self.view.sign_handler.turns:
-            [start_point, end_point, revers_start_point, revers_end_point] = self.calculation.calculation_four_dots(
-                turn)
-
-            temp_obj = {
-                "0": start_point, "1": start_point, "2": start_point,
-                "3": revers_end_point, "4": revers_end_point, "5": revers_start_point,
-                "5.1": revers_start_point, "6": revers_start_point,
-                "7": revers_end_point, "8": revers_end_point
-            }
-            grouped_objects = {}
-            for obj in turn.signs:
-                key = str(obj.number)
-                grouped_objects.setdefault(key, []).append(obj)
-
-            for key, items in grouped_objects.items():
-                x_current, y_current, x_prev, y_prev, azimuth = temp_obj[key]
-                x_current, y_current = self.converter.coordinateConverter(x_current, y_current, "epsg:4326",
-                                                                          "epsg:32635")
-                x_prev, y_prev = self.converter.coordinateConverter(x_prev, y_prev, "epsg:4326", "epsg:32635")
-
-                coefficient = 2
-                for item in items:
-                    item.azimuth = azimuth
-                    x1, y1, x2, y2 = self.calculation.calculate_result_line(item, coefficient,
-                                                                            x_prev, y_prev,
-                                                                            x_current, y_current)
-                    x1, y1 = self.converter.coordinateConverter(x1, y1, "epsg:32635", "epsg:4326")
-                    x2, y2 = self.converter.coordinateConverter(x2, y2, "epsg:32635", "epsg:4326")
-                    coefficient += 1
-                    feature = self.calculation.create_feature_object(x1, x2, y1, y2, item)
-                    features.append(feature)
-        return features
     def finish_processing(self):
-
         self.toggle_button_activity()
-        self.save_result()
+        self.finalHandler.save_result(self.view.sign_handler.result_signs, self.view.sign_handler.side_signs, self.view.sign_handler.turns)
+        #self.save_result()
         self.create_image_with_errors()
         self.view.cap.release()
         self.is_brake = True
         self.thread.join()
         self.thread = Thread(target=self.treatment, daemon=True)
         #cv2.destroyAllWindows()
-
        #self.stop_progressing()
 
-    def save_result(self):
-        path = config.PATH_TO_GEOJSON
 
-        features_signs = self.handling_signs()
-        features_turns = self.handling_turns()
-        features_side = self.handling_side()
-        features = features_signs + features_turns + features_side
-        signs_for_delete = []
-        for i in range(len(features)):
-            item_for_matching = features[i]
-            for j in range(len(features)):
-                item_with_which_matching = features[j]
-
-                item_with_x = item_with_which_matching["properties"]['w']
-                item_with_x = [int(re.findall(r'\b\d+\b', item)[0]) for item in item_with_x.split(',')]
-                average_item_with_x = sum(item_with_x) / len(item_with_x)
-                item_for_x = item_for_matching["properties"]['w']
-                item_for_x = [int(re.findall(r'\b\d+\b', item)[0]) for item in item_for_x.split(',')]
-                average_item_for_x = sum(item_for_x) / len(item_for_x)
-
-                item_for_car_x = item_for_matching["properties"]['car_coordinates_x']
-                item_for_car_x = [float(re.findall(r'\b\d+\b', item)[0]) for item in item_for_car_x.split(',')]
-                item_for_car_y = item_for_matching["properties"]['car_coordinates_y']
-                item_for_car_y = [float(re.findall(r'\b\d+\b', item)[0]) for item in item_for_car_y.split(',')]
-                item_with_car_x = item_with_which_matching["properties"]['car_coordinates_x']
-                item_with_car_x = [float(re.findall(r'\b\d+\b', item)[0]) for item in item_with_car_x.split(',')]
-                item_with_car_y = item_with_which_matching["properties"]['car_coordinates_y']
-                item_with_car_y = [float(re.findall(r'\b\d+\b', item)[0]) for item in item_with_car_y.split(',')]
-
-                if "del" not in features[i]:
-                    if average_item_for_x != average_item_with_x:
-                        if item_for_matching["properties"]["left"] == item_with_which_matching["properties"]["left"]:
-                            if item_for_matching["properties"]["type"] == item_with_which_matching["properties"][
-                                "type"]:
-                                distance = self.calculation.calculation_distance(item_for_car_x[-1],
-                                                                                 item_for_car_y[-1],
-                                                                                 item_with_car_x[-1],
-                                                                                 item_with_car_y[-1])
-                                if distance < 20:
-                                    item_for_azimuth = float(item_for_matching["properties"]["azimuth"])
-                                    item_with_azimuth = float(item_with_which_matching["properties"]["azimuth"])
-                                    difference_azimuth = abs(
-                                        self.calculation.calculate_azimuth_change(item_for_azimuth, item_with_azimuth))
-                                    if difference_azimuth < 30:
-                                        features[j]["del"] = 0
-                                        if int(features[j]["properties"]["length"]) > int(
-                                                features[i]["properties"]["length"]):
-                                            signs_for_delete.append(item_for_matching)
-                                        else:
-                                            signs_for_delete.append(item_with_which_matching)
-                                        break
-            #TODO Непонятно зачем но крашит программу
-            #for item in signs_for_delete:
-            #    print(item)
-            #    features.remove(item)
-        feature_collection = FeatureCollection(features)
-
-        with open(path, 'w', encoding='cp1251') as f:
-            dump(feature_collection, f, skipkeys=False, ensure_ascii=True)
-        print("save")
 
     def create_image_with_errors(self):
         result = []
